% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Germ.R
\name{Germ}
\alias{Germ}
\title{Bind two factors}
\usage{
Germ(city, thing)
}
\arguments{
\item{city}{factor}

\item{thing}{factor}
}
\value{
factor
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
\examples{
Germ(kin$pop[c(51, 51, 121)], job$av.g[c(51, 81, 23)])
}
